\name{bgbase_csv2ht}\alias{bgbase_csv2ht}\title{Convert the exported BG-BASE SPECIMENS csv to Darwin Core herbarium records template in CSV}\description{Convert the exported BG-BASE SPECIMENS CSV to Darwin Core herbarium records template in CSV.}\usage{bgbase_csv2ht(infile, outfile = NULL,     HERBARIUM = "Kadoorie Farm and Botanic Garden Herbarium (KFBG)",     HERBARIUM_CODE = "KFBG", TITLE = NA,     PROJECT = NA, PROCESSED_BY = NA, DATE_LASTMODIFIED = NA)}\arguments{  \item{infile}{The csv file exported from the SPECIMENS table by slist of BG-BASE}  \item{outfile}{The output csv file.}  \item{HERBARIUM}{Name of the herbarium}  \item{HERBARIUM_CODE}{Code of the herbarium}  \item{TITLE}{Plants of some place}  \item{PROJECT}{The project the specimen related with}  \item{PROCESSED_BY}{The name of the person that makes the conversion}  \item{DATE_LASTMODIFIED}{The date of the conversion}}\details{Many Botanic Gardens use the BG-BASE to manage the herbarium specimens. All the specimen records will go to the SPECIMENS table in BG-BASE. The records itself  has 771 fields in the SPECIMENS table, of which 191 are core fields, eg. Collector, Collector's Number. Exporting the records from BG-BASE specimens can be achieved by sending the command to slist interface. Here is an example of the command  exporting the records from BG-BASE slist:\code{ LIST SPECIMENS WITH RECORD_ID > 8000 and < 8546 ABUNDANCE 
ABUNDANCE_TNC ACCURACY ACC_NUM ACC_NUM_QUAL ALTITUDE ALTITUDE_QUAL 
ALTITUDE_UNIT ASPECT ASSOCIATED_INTCODE ASSOCIATED_INTERACTION 
ASSOCIATED_MAT ASSOCIATED_MAT_ACC_NUM ASSOCIATED_MAT_DEPO 
ASSOCIATED_MAT_DESITEM ASSOCIATED_MAT_IMAGE_NUM ASSOCIATED_MAT_MISC 
ASSOCIATED_MAT_NOMEN_TYPE_KIND ASSOCIATED_MAT_SPECIMEN_NUM 
ASSOCIATED_TAXA ASSOCIATED_TAXA_ABUNDANCE BARCODE BOX_NUM 
BRU_CODE CERTAINTY CITATION_DS CITATION_PAGE COLLECTED_WITH 
COLLECTION_MISC COLLECTOR COLL_DT COLL_DT_QUAL COLL_ID COLL_METHOD 
COLL_METHOD_MISC COLL_NAME COLL_NUM COLL_NUM_PREFIX_OLD 
COLL_NUM_SUFFIX_OLD COLL_SERIES COLL_TIME CONDITION 
COUNTRY_AS_GIVEN COUNTRY_CODE COUNTRY_ON_LABEL CULTIVATED 
CURATORIAL_NOTE DATE_AS_GIVEN DATE_ON_LABEL DEPOSITED DEPTH 
DEPTH_QUAL DEPTH_UNIT DESCRIPTION DESITEM DET_NUM DONOR FAMILY 
FILING_NAME_FREE FILING_NAME_NUM GARD_COLLECTION_MISC 
GARD_COLLECTOR GARD_COLL_DT GARD_COLL_DT_QUAL GARD_COLL_ID 
GARD_COLL_NUM GARD_DESCRIPTION GARD_LOCATION GENUS GEOG_AREA 
GEOG_REGION GEOLOGIC_AGE GEOLOGY GEOMORPHOLOGY GEOREF_SOURCE 
GEOREF_SOURCE_DS GPS_DATUM HABITAT HABITAT_TYPE HERB_COLLECTION 
HERB_NUM HYDROLOGY_CODE IMAGE_NUM_STORED INTRODUCED IS_TYPE 
KEYWORD KIND_OF_SPECIMEN LABEL_FOOTING LABEL_FORMAT LABEL_HEADING 
LABEL_IMAGE_NUM LABEL_PRINTED LABEL_STATUS LABEL_TEXT LABEL_VERIFIED 
LAND_OWNER LAT_DEGREE LAT_DIR LAT_MINUTE LAT_SECOND LINKED_WITH 
LOAN_ITEM LOCALITY LOCATION LOCATION_SENSITIVE LONG_DEGREE LONG_DIR 
LONG_MINUTE LONG_SECOND MISC MISC_INTERNAL NAME_FREE NAME_NUM 
NATIONAL_GRID NOMEN_TYPE_FREE NOMEN_TYPE_KIND NOMEN_TYPE_NOTE 
NOMEN_TYPE_NUM NUM_DUPLICATES NUM_IN_SET OCCUPANCY OCCUPANCY_UNIT 
OTHER_DATE OTHER_LABEL_INFO PACKET_CONTENT PHENOLOGY PHYTOREGION 
PLACE_NUM POPULATION_CENTER POPULATION_SIZE POPULATION_SIZE_QUAL 
PRECIPITATION PROJECTION PROJECT_CODE PROJECT_NUM PROTECTED_AREA 
PSOURCE PSOURCE_ACC_NUM PSOURCE_MISC PSOURCE_NUM RECORD_ID 
REC_UPDATEREC_UPD_BY REGION RESOLUTION RESTRICTION SAMPLE 
SECONDARY_ID SECOND_COLL_DT SEED_SOURCE SERIAL SEX SHEET_NUM 
SLOPE SOIL_TYPE SPECIMEN_LOCATION SPECIMEN_NUM SPECIMEN_SUB_LOCATION 
SPEC_CHAR SPEC_CHAR_NOTE STAGE SUBSTRATE SUB_CNT1 SUB_CNT2 SUB_CNT3 
TAXONOMIC_NOTE TENT_ID_MISC TREATMENT TREATMENT_MISC TREATMENT_TIME 
TREATMENT_UNIT TYPE_CITATION_DS TYPE_CITATION_PAGE USER1 USER10 USER2 
USER3 USER4 USER5 USER6 USER7 USER8 USER9 USES UTM VEGETATION 
VEGETATION_TYPE VERIFICATION_NUM VERI_INFO 
CSV c:\\BGBASE_EXPORT20150122.CSV by SPECIMEN_NUM }Note: None ASCII character will result in a corrupted file. The relationship between the fields in SPECIMENS table, the exported file from SPECIMENS table of BG-BASE and the Darwin Core herbarium records template could be found at the extdata folder of this package. The fields must be filled manually if corresponding field can not be found in the SPECIMENS table. (see the arguments).  }\value{A data frame}\references{BG-BASE \url{http://www.bg-base.com/}Relationship between the BG-BASE SPECIMENS TABLE and the herblabel template.xlsx}\author{Jinlong Zhang}\examples{bbsp_csv_file <- system.file("extdata", "BG_BASE_EXPORT20150121.CSV",                              package = "herblabel")bgbase_csv2ht(infile = bbsp_csv_file,  outfile = "herblabel.csv")}