\name{colcn2019}
\alias{colcn2019}
\docType{data}
\title{
Accepted Names of Catalogue of Life, China, 2019
}
\description{
The accepted names from the website Catalogue of Life, China, 2019
}
\usage{data("colcn2019")}
\format{
  A data frame with 40730 observations on the following 2 variables.
  \describe{
    \item{\code{NAME_CN}}{a character vector showing the Chinese Names of the species}
    \item{\code{SCIENTIFIC_NAME}}{a character vector showing the full scientific name of the species}
  }
}
\details{

Detailed description is available at Catalogue of Life China 2019 Annual Checklist (http://www.sp2000.org.cn/)

}
\source{
\url{http://www.chinaplantspecies.org/} accessed on 15 June 2019 and 
\url{http://www.sp2000.org.cn/} accessed on 15 June 2019
}

\examples{
data(colcn2019)
}
\keyword{datasets}
