\name{latin_source}
\alias{latin_source}
\docType{data}
\title{
Latin words
}
\description{
A string vector for the latin words used in botany. 
}
\usage{data("latin_source")}
\format{
  The format is:
 chr [1:185194] "aa" "aadenotrichopodus" "aageodonta" "aalius" "aambe" "aangelica" "aangustissima" ...
}
\details{
This checklist contains the latin words as well as all the genera for vascular plants and the specific epithet. All the words are in lower case. 
}
\source{
\url{ http://pan.baidu.com/share/link?shareid=172752&uk=2684982421 }
}
\references{
\url{ http://pan.baidu.com/share/link?shareid=172752&uk=2684982421 }
}
\examples{
data(latin_source)
}
\keyword{datasets}
