\name{spfoc}
\alias{spfoc}
\docType{data}
\title{
The Chinese Names from Flora of China
}
\description{
The dataframe contains Chinese names and the scientific names for the vascular plants, as well as all the Chinese Names of for the genera by Liu et al. 2014. 
}
\usage{data("spfoc")}
\format{
  A data frame with 74068 observations on the following 2 variables.
  \describe{
    \item{\code{NAME_CN}}{a character vector}
    \item{\code{SCIENTIFIC_NAME}}{a character vector}
  }
}
\details{
Encoding of the Chinese Characters is "UTF-8".
}
\source{
\url{ http://www.efloras.org/flora_page.aspx?flora_id=2 }
}
\references{
\url{ http://www.efloras.org/flora_page.aspx?flora_id=2 }
}
\examples{
data(spfoc)
}
\keyword{datasets}
