\name{tplsplist}
\alias{tplsplist}
\docType{data}
\title{
The Accepted Plant Names from the Plantlist
}
\description{
A character vector containing all the accepted names of higher plants from the Plant List Website.
}
\usage{data("tplsplist")}
\format{
  The format is:
 chr [1:403382] "Aegilotriticum erebunii (Gandilyan) van Slageren" ...
}
\details{
A character vector containing all the accepted names of higher plants from the Plant List Website.
}
\source{
The Plantlist Website \url{ http://www.theplantlist.org/1.1/browse/-/-/ }

The embbed database is available at \url{ http://pan.baidu.com/s/1hqHrW9I }
}
\references{

The Plant List (2013). Version 1.1. Published on the Internet; \url{ http://www.theplantlist.org/ } (accessed 26 July 2016). 

}
\examples{
data(tplsplist)
}
\keyword{datasets}
