\name{fill_sp_dwc}
\alias{fill_sp_dwc}
\title{
Fill in Darwin Core Species Information
}
\description{
Fill in Darwin Core Species Information
}
\usage{
fill_sp_dwc(dat)
}

\arguments{
  \item{dat}{
Darwin Core Herbarium Template
}
}
\details{
To be added
}
\value{
With Scientific Names Parsed
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Jinlong Zhang
}

\seealso{
See Also as \code{\link{herbarium_label}}
}
\examples{
### To be added.
}

\keyword{ Darwin Core }